+++ config/urls.py
@@ -21,6 +21,7 @@ from django.urls import include, path
 
 urlpatterns = [
     path('admin/', admin.site.urls),
+    path('api/', include('expenses.api.urls')),
     path('', include('expenses.urls')),
 ]